+++ codex_of_the_damned/static/js/toggleRulingForm.js
@@ -1,8 +1,0 @@
-function toggleRulingForm() {
-  let ruling = document.getElementById("ruling-form");
-  if (ruling.style.display == "none") {
-	ruling.style.display = "block";
-  } else {
-	ruling.style.display = "none";
-  }
-}