+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 1.11 (unreleased)
 -----------------
 
-- Nothing changed yet.
+- Use and display cards translations when on the non-english version of the site
 
 
 1.10 (2020-12-03)